+++ src/main/java/com/bildungsinsitut/deutschkurse/repository/KursRepository.java
@@ -10,10 +10,10 @@ import java.time.LocalDate;
 import java.util.List;
 
 @Repository
-public interface KursRepository extends JpaRepository<Kurs, Long> {
+public interface KursRepository extends JpaRepository<Kurs, Integer> {
     List<Kurs> findByStatus(KursStatusType status);
     List<Kurs> findByStatusIn(List<KursStatusType> statuses);
-    List<Kurs> findByTrainerId(Long trainerId);
+    List<Kurs> findByTrainerId(Integer trainerId);
     List<Kurs> findByStartdatumBetween(LocalDate start, LocalDate end);
 
     @Query("SELECT k FROM Kurs k WHERE k.aktuelleTeilnehmer < k.maxTeilnehmer AND k.status = 'geplant'")
